pyproject.toml
[tool.poetry]
name = "copy-channel-telegram-bot"
version = "0.1.0"
description = "Бот который перекидывать сообщения из одного канала в другой"
authors = ["wselfjes <wselfjes@gmail.com>"]
readme = "README.md"
packages = [{include = "copy_channel_telegram_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.3"
pydantic = "^1.10.7"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
